// src/services/liteCache/liteCache.interface.ts
export type AppSchema = {
  cache: Record<string, any>;
  jobs: JobEntry[];
};

export interface JobEntry {
  id: string;
  schedule: string;
  isRepeating: boolean;
  refTable?: "action" | "device" | "notification" | "command";
  refId?: string;
  command?: {
    deviceId?: string;
    key: string;
    value?: string | number | boolean;
  };
}

// export type DataSchema = {
//   jobs: JobEntry[];
// };